+++ include/common.h
@@ -27,15 +27,15 @@
 #if defined(CTR_NATIVE)
 static inline void *CTR_PsyqMemmove(void *dest, const void *src, s32 count)
 {
-	// NOTE(aalhendi): Retail PSYQ memmove at NTSC-U 926 0x80077e38 returns
-	// immediately for signed lengths <= 0. Host libc takes size_t, so native
-	// must preserve the signed PSYQ contract for ASM-verified game code.
+	// NOTE(aalhendi): Retail PSYQ memmove at 0x80077e38 returns
+	// immediately for signed lengths <= 0. Host libc takes u32, so native
+	// must preserve the signed PSYQ contract for retail-shaped game code.
 	if (count <= 0)
 	{
 		return dest;
 	}
 
-	return memmove(dest, src, (size_t)count);
+	return memmove(dest, src, (u32)count);
 }
 
 #define memmove(dest, src, count) CTR_PsyqMemmove((dest), (src), (s32)(count))
